Basic windows port
